@cutierobot profile page new style design tempalte
After consolting with @aXises  today in class a slightly different
design for the profile page was created. The detials text will now
appear in the orange header.

The template is practically done jsut a bit more style things to do like
aligning stuff. I did some floating so if anything breaks or goes where
it is not meant to its probably related to those fixes.